+++ 阶加.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main() {
-	unsigned long x;
+	unsigned long long x;
 	fprintf(stderr,"Please type a num: "); 
-	scanf("%lu",&x);
+	scanf("%llu",&x);
 	fprintf(stderr,"result: ");
 	if(x%2==0)
-		fprintf(stdout,"%lu\n",(x+1)*(x/2));
+		fprintf(stdout,"%llu\n",(x+1)*(x/2));
 	else
-		fprintf(stdout,"%lu\n",x*(x/2)+x);
+		fprintf(stdout,"%llu\n",x*(x/2)+x);
 	return 0;
 }
